+++ src/core/settings.h
@@ -156,7 +156,7 @@ struct Values {
     std::string sink_id;
     bool enable_audio_stretching;
     std::string audio_device_id;
-    u16 sound_volume;
+    u8 sound_volume;
 
     // Camera
     std::array<std::string, Service::CAM::NumCameras> camera_name;
